Implemented the CRUD operations for the User Repository/Service/Controller. Added a helpfull UserRequestDTO, that might be used to easier transffer data between operations. Added a UserControllerAdvice, in order to communicate better messages in case of specific errors that might be thrown.

+++ alpr-server/src/main/java/isd/alprserver/service/UserService.java
@@ -44,7 +44,7 @@ public class UserService implements UserDetailsService {
         return true;
     }
 
-    public boolean deleteUser(int userId) {
+    public boolean deleteUser(long userId) {
         if(userRepository.findById(userId).isPresent()) {
             userRepository.deleteById(userId);
             return true;
